Use MYSQL_LONGLONG type in getUID query
The primary key of the usr table is a BIGINT, so the MySQL type used to
fetch the id should be 8 bytes-long. A long long guarantees such size.

The problem to fix here is not much the chance of overflowing the value,
but the wrong representation of the ID in the VOMS server logs.

Issue: VOMS-748